requirements.txt
#Python libraries required for this project

requests
psycopg2-binary
python-dotenv
Flask